requirements.txt
langchain
langchain-google-genai
google-generativeai
requests
python-dotenv